Rename math.h to avoid conflict with cmath header
