+++ src/lemon/binomial_heap.h
@@ -26,7 +26,7 @@
 #include <vector>
 #include <utility>
 #include <functional>
-#include <lemon/math.h>
+#include <lemon/lmath.h>
 #include <lemon/counter.h>
 
 namespace lemon {
