+++ src/lemon/color.h
@@ -20,7 +20,7 @@
 #define LEMON_COLOR_H
 
 #include<vector>
-#include<lemon/math.h>
+#include<lemon/lmath.h>
 #include<lemon/maps.h>
